🛠️refactor: 코드리뷰 피드백 반영
- 캘린더 스타일 수정
- 디렉토리 명 수정
- 스토리북 수정
