+++ package.json
@@ -38,7 +38,7 @@
     "@storybook/addon-onboarding": "^8.0.9",
     "@storybook/blocks": "^8.0.9",
     "@storybook/nextjs": "^8.0.9",
-    "@storybook/react": "^8.0.9",
+    "@storybook/react": "^8.1.5",
     "@storybook/test": "^8.0.9",
     "@tanstack/react-query-devtools": "^5.32.0",
     "@trivago/prettier-plugin-sort-imports": "^4.3.0",
