+++ src/app/_components/client/Calender/FeedCalender/components/index.ts
@@ -1,3 +1,3 @@
 export { default as DailtyDDuDu } from "./DailyDDuDu/DailyDDuDu";
-export { default as FeedCalenderDayContent } from "./useFeedCalenderDayConetent/FeedCalenderDayContent";
+export { default as FeedCalenderDayContent } from "./FeedCalenderDayConetent/FeedCalenderDayContent";
 export { default as FeedCalenderHeader } from "./FeedCalenderHeader/FeedCalenderHeader";
